Fix error

+++ src/intel_8253_pit.c
@@ -13,7 +13,7 @@ void drivers_intel_8253_pit_initialize(const unsigned int freq)
     const uint8_t l = divisor & 0xff;
     const uint8_t h = (divisor >> 8) & 0xff;
 
-    __asm__ __volatile__("outb %1, %0" : : "dN" (0x43), "a" (0x36));
+    __asm__ __volatile__("outb %1, %0" : : "dN" (0x43), "a" ((uint8_t)0x36));
     __asm__ __volatile__("outb %1, %0" : : "dN" (0x40), "a" (l));
     __asm__ __volatile__("outb %1, %0" : : "dN" (0x40), "a" (h));
 }
