+++ src/shutdown.c
@@ -4,6 +4,8 @@
 
 #include <drivers/shutdown.h>
 
+#include <stdint.h>
+
 void drivers_shutdown_halt()
 {
 #ifdef ASM_X86
@@ -19,9 +21,9 @@ void drivers_shutdown_poweroff()
 {
 #ifdef ASM_X86
     // QEMU >= 2.0
-    __asm__ __volatile__("outw %1, %0" : : "dN" (0x604), "a" (0x2000));
+    __asm__ __volatile__("outw %1, %0" : : "dN" (0x604), "a" ((uint16_t)0x2000));
     // QEMU < 2.0
-    __asm__ __volatile__("outw %1, %0" : : "dN" (0xB004), "a" (0x2000));
+    __asm__ __volatile__("outw %1, %0" : : "dN" (0xB004), "a" ((uint16_t)0x2000));
 #endif
 
     // If we can't poweroff then we halt
